+++ src/consumer/component/consumer-component.js
@@ -2,7 +2,7 @@ import path from 'path';
 import fs from 'fs';
 import { mkdirp, isString } from '../../utils';
 import BitJson from '../bit-json';
-import { Impl, Specs, Dist, License } from '../component/sources';
+import { Impl, Specs, Dist, License, Misc } from '../component/sources';
 import ConsumerBitJson from '../bit-json/consumer-bit-json';
 import Consumer from '../consumer';
 import BitId from '../../bit-id/bit-id';
@@ -30,12 +30,14 @@ export type ComponentProps = {
   scope?: ?string,
   implFile?: ?string,
   specsFile?: ?string,
+  miscFiles?: ?string[],
   compilerId?: ?BitId,
   testerId?: ?BitId,
   dependencies?: ?BitIds,
   packageDependencies?: ?Object,
   impl?: ?Impl|string,
   specs?: ?Specs|string,
+  misc?: ?Misc|[],
   docs?: ?Doclet[],
   dist?: ?Dist,
   specsResults?: ?SpecsResults,
@@ -49,6 +51,7 @@ export default class Component {
   scope: ?string;
   implFile: string;
   specsFile: string;
+  miscFiles: string[];
   compilerId: ?BitId;
   testerId: ?BitId;
   dependencies: BitIds;
@@ -56,6 +59,7 @@ export default class Component {
   _impl: ?Impl|string;
   _specs: ?Specs|string;
   _docs: ?Doclet[];
+  _misc: ?Misc|[];
   dist: ?Dist;
   specsResults: ?SpecsResults;
   license: ?License;
@@ -84,6 +88,20 @@ export default class Component {
     return this._specs;
   }
 
+  set misc(val: Misc) { this._misc = val; }
+
+  get misc(): ?Misc {
+    if (!this._misc) return null;
+    if (this._misc instanceof Misc) return this._misc;
+
+    if (Array.isArray(this._misc)) {
+      // $FlowFixMe
+      this._misc = Misc.load(this._misc);
+    }
+    // $FlowFixMe
+    return this._misc;
+  }
+
   get id(): BitId {
     if (!this.scope || !this.version) {
       console.error(this);
@@ -110,12 +128,14 @@ export default class Component {
     scope,
     implFile,
     specsFile,
+    miscFiles,
     compilerId,
     testerId,
     dependencies,
     packageDependencies,
     impl,
     specs,
+    misc,
     docs,
     dist,
     specsResults,
@@ -127,12 +147,14 @@ export default class Component {
     this.scope = scope;
     this.implFile = implFile || DEFAULT_IMPL_NAME;
     this.specsFile = specsFile || DEFAULT_SPECS_NAME;
+    this.miscFiles = miscFiles || [];
     this.compilerId = compilerId;
     this.testerId = testerId;
     this.dependencies = dependencies || new BitIds();
     this.packageDependencies = packageDependencies || {};
     this._specs = specs;
     this._impl = impl;
+    this._misc = misc;
     this._docs = docs;
     this.dist = dist;
     this.specsResults = specsResults;
@@ -145,6 +167,7 @@ export default class Component {
       scope: this.scope,
       impl: this.implFile,
       spec: this.specsFile,
+      miscFiles: this.miscFiles,
       compiler: this.compilerId ? this.compilerId.toString() : NO_PLUGIN_TYPE,
       tester: this.testerId ? this.testerId.toString() : NO_PLUGIN_TYPE,
       dependencies: this.dependencies.toObject(),
@@ -160,6 +183,7 @@ export default class Component {
     return mkdirp(bitDir)
       .then(() => this.impl.write(bitDir, this.implFile, force))
       .then(() => { return this.specs ? this.specs.write(bitDir, this.specsFile, force) : undefined; })
+      .then(() => { return this.misc ? this.misc.write(bitDir, this.miscFiles, force) : undefined; })
       .then(() => { return this.dist ? this.dist.write(bitDir, this.implFile, force) : undefined; })
       .then(() => { return withBitJson ? this.writeBitJson(bitDir, force): undefined; })
       .then(() => { return this.license && this.license.src ? this.license.write(bitDir, force) : undefined; })
@@ -342,12 +366,14 @@ export default class Component {
         box: path.basename(path.dirname(bitDir)),
         implFile: bitJson.getImplBasename(),
         specsFile: bitJson.getSpecBasename(),
+        miscFiles: bitJson.getMiscFiles(),
         compilerId: BitId.parse(bitJson.compilerId),
         testerId: BitId.parse(bitJson.testerId),
         dependencies: BitIds.fromObject(bitJson.dependencies),
         packageDependencies: bitJson.packageDependencies,
         impl: path.join(bitDir, bitJson.getImplBasename()),
         specs: path.join(bitDir, bitJson.getSpecBasename()),
+        misc: bitJson.getMiscFiles().map(misc => path.join(bitDir, misc)),
       });
     });
   }
@@ -361,6 +387,7 @@ export default class Component {
   }, scope: Scope) {
     const implFile = consumerBitJson.getImplBasename();
     const specsFile = consumerBitJson.getSpecBasename();
+    const miscFiles = consumerBitJson.getMiscFiles();
     const compilerId = BitId.parse(consumerBitJson.compilerId);
     const testerId = BitId.parse(consumerBitJson.testerId);
 
@@ -371,6 +398,7 @@ export default class Component {
       scope: scopeName,
       implFile,
       specsFile,
+      miscFiles,
       compilerId,
       testerId,
       impl: Impl.create(name, compilerId, scope),