+++ src/consumer/component/sources/index.js
@@ -3,6 +3,7 @@ import Specs from './specs';
 import Source from './source';
 import Dist from './dist';
 import License from './license';
+import Misc from './misc';
 
 export {
   Impl,
@@ -10,4 +11,5 @@ export {
   Source,
   Dist,
   License,
+  Misc,
 };